pyproject.toml
[tool.poetry]
name = "django-2"
version = "0.1.0"
description = ""
authors = ["Isfireomat <dfdfgfgf666@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
gunicorn = "^23.0.0"
djangorestframework = "^3.15.2"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
psycopg2 = "^2.9.10"
redis = "^5.2.0"
celery = {extras = ["redis"], version = "^5.4.0"}
drf-spectacular = "^0.27.2"
django-celery-beat = "^2.7.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
